requirements.txt
# python==3.8.12
biopython==1.79
dgl==0.7.2
joblib==1.1.0
logzero==1.7.0
numpy==1.21.5
pandas==1.3.5
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
torch==1.10.0
tqdm==4.62.3
diamond==2.0.14
cudatoolkit==11.3.1
